datastore-postgres: hoist test database test fixture from postgres source and destination
