+++ airbyte-cdk/java/airbyte-cdk/datastore-postgres/build.gradle
@@ -11,5 +11,7 @@ dependencies {
 
     api 'org.postgresql:postgresql:42.6.0'
 
+    testFixturesImplementation testFixtures(project(':airbyte-cdk:java:airbyte-cdk:core'))
+
     testFixturesApi 'org.testcontainers:postgresql:1.19.0'
 }
