+++ airbyte-integrations/connectors/destination-postgres/src/test-integration/java/io/airbyte/integrations/destination/postgres/typing_deduping/PostgresTypingDedupingTest.java
@@ -7,7 +7,7 @@ package io.airbyte.integrations.destination.postgres.typing_deduping;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.airbyte.integrations.destination.postgres.PostgresDestination;
-import io.airbyte.integrations.destination.postgres.PostgresTestDatabase;
+import io.airbyte.integrations.postgres.PostgresTestDatabase;
 import javax.sql.DataSource;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
