+++ airbyte-integrations/connectors/source-postgres/build.gradle
@@ -27,8 +27,6 @@ dependencies {
     implementation 'io.debezium:debezium-embedded:2.4.0.Final'
     implementation 'io.debezium:debezium-connector-postgres:2.4.0.Final'
 
-    testFixturesApi 'org.testcontainers:postgresql:1.19.0'
-
     testImplementation 'org.hamcrest:hamcrest-all:1.3'
 }
